// src/shared/enums/error-messages.enum.ts
export enum ErrorMessages {
  ZOD_VALIDATION = 'Erro ao validar os dados.',
  PERSON_NOT_FOUND = 'Pessoa não encontrada.',
  PERSON_CPF_ALREADY_REGISTERED = 'Pessoa com esse CPF já está cadastrada.',
  PERSON_EMAIL_ALREADY_REGISTERED = 'Pessoa com esse e-mail já está cadastrada.',
  USER_EMAIL_ALREADY_REGISTERED = 'Usuário com esse e-mail já está cadastrado.',
  USER_NOT_FOUND = 'Usuário não encontrado.',
  LOGIN_ERROR = 'Não foi possível realizar o login.',
}